+++ packetizers/arq/ARQ.cpp
@@ -234,6 +234,7 @@ inline void ARQ::process_packet(RFMessage &m) {
 #ifdef USE_PML
 	PML->packet_in(m);
 #endif
+	this->trigger_led();
 
 
 	uint8_t id, seg;
